scripts/requirements.txt
requests>=2.31.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
anthropic>=0.34.0
sendgrid>=6.11.0
